+++ src/objects/Address.py
@@ -1,0 +1,4 @@
+from objects.Field import Field
+
+class Address(Field):
+    pass
