+++ src/objects/Field.py
@@ -1,0 +1,4 @@
+
+class Field:
+    def __init__(self, value):
+        self.value = value
