+++ src/objects/Name.py
@@ -1,0 +1,4 @@
+from objects.Field import Field
+
+class Name(Field):
+    pass
