+++ src/objects/Phone.py
@@ -1,0 +1,10 @@
+from objects.Field import Field
+
+class Phone(Field):
+    def __init__(self, value):
+        super().__init__(value)
+        self.validate()
+
+    def validate(self):
+        if not self.value.isdigit() or len(self.value) != 10:
+            raise ValueError("Phone number must consist of 10 digits.")
